# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CED OPS API",
    "description": "REST API for the CED (Continuous Electro-Deposition) Operations management system. Configuration-driven report engine over a module hierarchy (module types, modules, template versions, processes, process parameters and global parameters), report sessions and completed reports, plus dashboard, unified search, analytics and role-based access control.",
    "termsOfService": "https://aerotech.com/terms",
    "contact": {
      "name": "AeroTech Support",
      "url": "https://aerotech.com",
      "email": "support@aerotech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://aerotech.com/license"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Default Server URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/report-engine/start": {
      "post": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN, ADMIN, OPERATOR",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReportSessionResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReportSessionResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/{sessionId}/current": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReportProcessStep"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/{sessionId}/save-next": {
      "post": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN, ADMIN, OPERATOR",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordProcessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RecordProcessResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/{sessionId}/save-submit": {
      "post": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN, ADMIN, OPERATOR",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordProcessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RecordProcessResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/{sessionId}/recorded": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecordedProcessItem"
                      }
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/reports/{reportId}": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CompletedReportResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/reports/my": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CompletedReportResponse"
                      }
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/report-engine/sessions/my": {
      "get": {
        "tags": [
          "Report Engine (Module-driven)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReportSessionResponse"
                      }
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "StartReportRequest": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "integer",
            "format": "int64"
          },
          "shiftId": {
            "type": "integer",
            "format": "int64"
          },
          "lineId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "RecordedValueRequest": {
        "type": "object",
        "properties": {
          "processParameterId": {
            "type": "integer",
            "format": "int64"
          },
          "observedValue": {
            "type": "string"
          }
        }
      },
      "ProcessParameterField": {
        "type": "object",
        "properties": {
          "processParameterId": {
            "type": "integer",
            "format": "int64"
          },
          "parameterId": {
            "type": "integer",
            "format": "int64"
          },
          "parameterName": {
            "type": "string"
          },
          "inputType": {
            "type": "string",
            "enum": [
              "NUMBER",
              "// Numeric input\n    TEXT",
              "// Text input\n    BOOLEAN",
              "// Boolean toggle input\n    DROPDOWN // Dropdown selection input"
            ]
          },
          "mandatory": {
            "type": "boolean"
          },
          "unit": {
            "type": "string"
          },
          "minimumValue": {
            "type": "number"
          },
          "maximumValue": {
            "type": "number"
          },
          "defaultValue": {
            "type": "string"
          }
        }
      },
      "RecordedProcessItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processId": {
            "type": "integer",
            "format": "int64"
          },
          "processName": {
            "type": "string"
          },
          "processOrderSnapshot": {
            "type": "integer",
            "format": "int32"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordedValueItem"
            }
          }
        }
      },
      "ReportSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "moduleId": {
            "type": "integer",
            "format": "int64"
          },
          "moduleName": {
            "type": "string"
          },
          "templateVersionId": {
            "type": "integer",
            "format": "int64"
          },
          "versionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "currentProcessId": {
            "type": "integer",
            "format": "int64"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedProcessCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED"
            ]
          }
        }
      },
      "ReportProcessStep": {
        "type": "object",
        "properties": {
          "processId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "lastProcess": {
            "type": "boolean"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessParameterField"
            }
          }
        }
      },
      "RecordProcessRequest": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordedValueRequest"
            }
          }
        }
      },
      "RecordProcessResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int64"
          },
          "completedProcessCount": {
            "type": "integer",
            "format": "int32"
          },
          "nextProcess": {
            "$ref": "#/components/schemas/ReportProcessStep"
          },
          "reportCompleted": {
            "type": "boolean"
          },
          "report": {
            "$ref": "#/components/schemas/CompletedReportResponse"
          }
        }
      },
      "RecordedValueItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processParameterId": {
            "type": "integer",
            "format": "int64"
          },
          "parameterId": {
            "type": "integer",
            "format": "int64"
          },
          "parameterName": {
            "type": "string"
          },
          "inputType": {
            "type": "string",
            "enum": [
              "NUMBER",
              "// Numeric input\n    TEXT",
              "// Text input\n    BOOLEAN",
              "// Boolean toggle input\n    DROPDOWN // Dropdown selection input"
            ]
          },
          "observedValue": {
            "type": "string"
          }
        }
      },
      "CompletedReportResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reportNumber": {
            "type": "string"
          },
          "moduleName": {
            "type": "string"
          },
          "versionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "prefix": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "// Report has been saved as draft and not yet submitted\n    SUBMITTED",
              "// Report has been submitted for approval\n    APPROVED",
              "// Report has been approved\n    REJECTED // Report has been rejected"
            ]
          },
          "sessionId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}